feat: add breaking change footer completion
This change refactors the `get_completions` method a little to simplify
the logic for when to get completions for conventional commits.

Furthermore it adds an additonal completion for the rest of the buffer
to allow for a 'BREAKING CHANGE' footer.
The completion will also auto prepend '!' before the colon in the header
if its missing.